pyproject.toml
[project]
name = "wire"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "feedparser>=6.0.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10.0",
    "scikit-learn>=1.6.0",
    "anthropic>=0.43.0",
    "jinja2>=3.1.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
